updated versioning, added a few unit tests and removed old Python unit tests
