+++ .ci/Datasets/Compute/SimulationId.json
@@ -1,0 +1,1 @@
+{ "_t" : "BH.oM.Data.Library.Dataset", "BHoM_Guid" : "eea548b4-5137-46ab-afaf-b5bcb86ff02a", "Name" : "SimulationId", "Fragments" : [], "Tags" : [], "CustomData" : { }, "SourceInformation" : { "_t" : "BH.oM.Data.Library.Source", "BHoM_Guid" : "4021e615-35a5-4f50-9bca-bbf26d733986", "Name" : "", "Fragments" : [], "Tags" : [], "CustomData" : { }, "SourceLink" : "", "Title" : "SimulationId", "Author" : "tgerrish", "ItemReference" : "", "Version" : "", "Publisher" : "", "Schema" : "", "Language" : "", "Location" : "", "Copyright" : "", "Contributors" : "", "Confidence" : "Undefined" }, "TimeOfCreation" : { "$date" : 1669726598781 }, "Data" : [{ "_t" : "BH.oM.Test.UnitTests.UnitTest", "BHoM_Guid" : "17d09334-c3b9-4415-96f2-a37f217f7c6f", "Name" : "", "Fragments" : [], "Tags" : [], "CustomData" : { }, "Method" : { "_t" : "System.Reflection.MethodBase", "TypeName" : "{ \"_t\" : \"System.Type\", \"Name\" : \"BH.Engine.LadybugTools.Compute, LadybugTools_Engine, Version=6.0.0.0, Culture=neutral, PublicKeyToken=null\", \"_bhomVersion\" : \"5.3\" }", "MethodName" : "SimulationId", "Parameters" : ["{ \"_t\" : \"System.Type\", \"Name\" : \"System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089\", \"_bhomVersion\" : \"5.3\" }", "{ \"_t\" : \"System.Type\", \"Name\" : \"BH.oM.LadybugTools.ILBTMaterial\", \"_bhomVersion\" : \"5.3\" }", "{ \"_t\" : \"System.Type\", \"Name\" : \"BH.oM.LadybugTools.ILBTMaterial\", \"_bhomVersion\" : \"5.3\" }"], "_bhomVersion" : "5.3" }, "Data" : [{ "_t" : "BH.oM.Test.UnitTests.TestData", "BHoM_Guid" : "f1709cf4-ed0d-470a-bc2c-b4f8833672e5", "Name" : "", "Fragments" : [], "Tags" : [], "CustomData" : { }, "Inputs" : ["C:/example.csv", { "_t" : "BH.oM.LadybugTools.OpaqueMaterial", "Identifier" : "Material%^ ]@A", "Roughness" : "MediumRough", "Thickness" : 0.10000000000000001, "Conductivity" : 2.0, "Density" : 2000.0, "SpecificHeat" : 800.0, "ThermalAbsorptance" : 0.90000000000000002, "SolarAbsorptance" : 0.69999999999999996, "VisibleAbsorptance" : 0.69999999999999996 }, { "_t" : "BH.oM.LadybugTools.OpaqueMaterial", "Identifier" : "Material%^ ]@A", "Roughness" : "MediumRough", "Thickness" : 0.10000000000000001, "Conductivity" : 2.0, "Density" : 2000.0, "SpecificHeat" : 800.0, "ThermalAbsorptance" : 0.90000000000000002, "SolarAbsorptance" : 0.69999999999999996, "VisibleAbsorptance" : 0.69999999999999996 }], "Outputs" : ["example__MaterialA__MaterialA"] }] }], "_bhomVersion" : "5.3" }
