+++ .ci/Datasets/Query/GetMaterial.json
@@ -1,0 +1,1 @@
+{ "_t" : "BH.oM.Data.Library.Dataset", "BHoM_Guid" : "d46508e1-6124-4547-9a0a-62fe4cb9807d", "Name" : "GetMaterial", "Fragments" : [], "Tags" : [], "CustomData" : { }, "SourceInformation" : { "_t" : "BH.oM.Data.Library.Source", "BHoM_Guid" : "93b93ea8-d698-434b-97fc-13c862a51199", "Name" : "", "Fragments" : [], "Tags" : [], "CustomData" : { }, "SourceLink" : "", "Title" : "GetMaterial", "Author" : "tgerrish", "ItemReference" : "", "Version" : "", "Publisher" : "", "Schema" : "", "Language" : "", "Location" : "", "Copyright" : "", "Contributors" : "", "Confidence" : "Undefined" }, "TimeOfCreation" : { "$date" : 1669726598781 }, "Data" : [{ "_t" : "BH.oM.Test.UnitTests.UnitTest", "BHoM_Guid" : "1aa9afcd-d90b-4176-8bb9-8040a8c67220", "Name" : "", "Fragments" : [], "Tags" : [], "CustomData" : { }, "Method" : { "_t" : "System.Reflection.MethodBase", "TypeName" : "{ \"_t\" : \"System.Type\", \"Name\" : \"BH.Engine.LadybugTools.Query, LadybugTools_Engine, Version=6.0.0.0, Culture=neutral, PublicKeyToken=null\", \"_bhomVersion\" : \"5.3\" }", "MethodName" : "GetMaterial", "Parameters" : ["{ \"_t\" : \"System.Type\", \"Name\" : \"BH.oM.LadybugTools.Materials\", \"_bhomVersion\" : \"5.3\" }"], "_bhomVersion" : "5.3" }, "Data" : [{ "_t" : "BH.oM.Test.UnitTests.TestData", "BHoM_Guid" : "270517ad-f806-400b-bb08-6116bc78b16d", "Name" : "", "Fragments" : [], "Tags" : [], "CustomData" : { }, "Inputs" : [2], "Outputs" : [{ "_t" : "BH.oM.LadybugTools.OpaqueMaterial", "Identifier" : "Concrete Pavement", "Roughness" : "MediumRough", "Thickness" : 0.20000000000000001, "Conductivity" : 1.73, "Density" : 2243.0, "SpecificHeat" : 837.0, "ThermalAbsorptance" : 0.90000000000000002, "SolarAbsorptance" : 0.65000000000000002, "VisibleAbsorptance" : 0.65000000000000002 }] }] }], "_bhomVersion" : "5.3" }
