+++ .ci/Datasets/Query/GetTypology.json
@@ -1,0 +1,1 @@
+{ "_t" : "BH.oM.Data.Library.Dataset", "BHoM_Guid" : "949240d2-6c81-4a3f-99cf-3a05b59908f5", "Name" : "GetTypology", "Fragments" : [], "Tags" : [], "CustomData" : { }, "SourceInformation" : { "_t" : "BH.oM.Data.Library.Source", "BHoM_Guid" : "9a6e5fb8-cf78-40cd-bd4d-008baec29c9d", "Name" : "", "Fragments" : [], "Tags" : [], "CustomData" : { }, "SourceLink" : "", "Title" : "GetTypology", "Author" : "tgerrish", "ItemReference" : "", "Version" : "", "Publisher" : "", "Schema" : "", "Language" : "", "Location" : "", "Copyright" : "", "Contributors" : "", "Confidence" : "Undefined" }, "TimeOfCreation" : { "$date" : 1669726598781 }, "Data" : [{ "_t" : "BH.oM.Test.UnitTests.UnitTest", "BHoM_Guid" : "2a953545-c6c9-4e61-8033-0c6ca84abaea", "Name" : "", "Fragments" : [], "Tags" : [], "CustomData" : { }, "Method" : { "_t" : "System.Reflection.MethodBase", "TypeName" : "{ \"_t\" : \"System.Type\", \"Name\" : \"BH.Engine.LadybugTools.Query, LadybugTools_Engine, Version=6.0.0.0, Culture=neutral, PublicKeyToken=null\", \"_bhomVersion\" : \"5.3\" }", "MethodName" : "GetTypology", "Parameters" : ["{ \"_t\" : \"System.Type\", \"Name\" : \"BH.oM.LadybugTools.Typologies\", \"_bhomVersion\" : \"5.3\" }"], "_bhomVersion" : "5.3" }, "Data" : [{ "_t" : "BH.oM.Test.UnitTests.TestData", "BHoM_Guid" : "1159946c-79e4-435c-97a9-c95a1de978f1", "Name" : "", "Fragments" : [], "Tags" : [], "CustomData" : { }, "Inputs" : [25], "Outputs" : [{ "_t" : "BH.oM.LadybugTools.Typology", "BHoM_Guid" : "04b411ae-05c5-4b43-8606-f4f0c8ce5950", "Name" : "East-west shelter", "Fragments" : [], "Tags" : [], "CustomData" : { }, "Shelters" : [{ "_t" : "BH.oM.LadybugTools.Shelter", "BHoM_Guid" : "d0051372-7e1c-4533-a2e0-ae6241c2d9fd", "Name" : "", "Fragments" : [], "Tags" : [], "CustomData" : { }, "WindPorosity" : 0.0, "RadiationPorosity" : 0.0, "AzimuthRange" : [67.5, 112.5], "AltitudeRange" : [0.0, 70.0] }, { "_t" : "BH.oM.LadybugTools.Shelter", "BHoM_Guid" : "c5e2d6e6-ac67-4228-9db2-e5b22b74730a", "Name" : "", "Fragments" : [], "Tags" : [], "CustomData" : { }, "WindPorosity" : 0.0, "RadiationPorosity" : 0.0, "AzimuthRange" : [247.5, 292.5], "AltitudeRange" : [0.0, 70.0] }], "EvaporativeCoolingEffectiveness" : 0.0, "WindSpeedAdjustment" : 1.0 }] }] }], "_bhomVersion" : "5.3" }
