--- .ci/unit_tests/python/files/example.pts
@@ -1,100 +1,0 @@
--4.5 -4.5 0.800000011920929 0.0 0.0 1.0
--4.5 -3.5 0.800000011920929 0.0 0.0 1.0
--4.5 -2.5 0.800000011920929 0.0 0.0 1.0
--4.5 -1.5 0.800000011920929 0.0 0.0 1.0
--4.5 -0.5 0.800000011920929 0.0 0.0 1.0
--4.5 0.5 0.800000011920929 0.0 0.0 1.0
--4.5 1.5 0.800000011920929 0.0 0.0 1.0
--4.5 2.5 0.800000011920929 0.0 0.0 1.0
--4.5 3.5 0.800000011920929 0.0 0.0 1.0
--4.5 4.5 0.800000011920929 0.0 0.0 1.0
--3.5 -4.5 0.800000011920929 0.0 0.0 1.0
--3.5 -3.5 0.800000011920929 0.0 0.0 1.0
--3.5 -2.5 0.800000011920929 0.0 0.0 1.0
--3.5 -1.5 0.800000011920929 0.0 0.0 1.0
--3.5 -0.5 0.800000011920929 0.0 0.0 1.0
--3.5 0.5 0.800000011920929 0.0 0.0 1.0
--3.5 1.5 0.800000011920929 0.0 0.0 1.0
--3.5 2.5 0.800000011920929 0.0 0.0 1.0
--3.5 3.5 0.800000011920929 0.0 0.0 1.0
--3.5 4.5 0.800000011920929 0.0 0.0 1.0
--2.5 -4.5 0.800000011920929 0.0 0.0 1.0
--2.5 -3.5 0.800000011920929 0.0 0.0 1.0
--2.5 -2.5 0.800000011920929 0.0 0.0 1.0
--2.5 -1.5 0.800000011920929 0.0 0.0 1.0
--2.5 -0.5 0.800000011920929 0.0 0.0 1.0
--2.5 0.5 0.800000011920929 0.0 0.0 1.0
--2.5 1.5 0.800000011920929 0.0 0.0 1.0
--2.5 2.5 0.800000011920929 0.0 0.0 1.0
--2.5 3.5 0.800000011920929 0.0 0.0 1.0
--2.5 4.5 0.800000011920929 0.0 0.0 1.0
--1.5 -4.5 0.800000011920929 0.0 0.0 1.0
--1.5 -3.5 0.800000011920929 0.0 0.0 1.0
--1.5 -2.5 0.800000011920929 0.0 0.0 1.0
--1.5 -1.5 0.800000011920929 0.0 0.0 1.0
--1.5 -0.5 0.800000011920929 0.0 0.0 1.0
--1.5 0.5 0.800000011920929 0.0 0.0 1.0
--1.5 1.5 0.800000011920929 0.0 0.0 1.0
--1.5 2.5 0.800000011920929 0.0 0.0 1.0
--1.5 3.5 0.800000011920929 0.0 0.0 1.0
--1.5 4.5 0.800000011920929 0.0 0.0 1.0
--0.5 -4.5 0.800000011920929 0.0 0.0 1.0
--0.5 -3.5 0.800000011920929 0.0 0.0 1.0
--0.4999999999999999 -2.5 0.800000011920929 0.0 0.0 1.0
--0.5 -1.5 0.800000011920929 0.0 0.0 1.0
--0.5 -0.5 0.8000000119209291 0.0 0.0 1.0
--0.5 0.5 0.800000011920929 0.0 0.0 1.0
--0.5 1.5000000000000002 0.8000000119209291 0.0 0.0 1.0
--0.5 2.5 0.8000000119209291 0.0 0.0 1.0
--0.5 3.5 0.800000011920929 0.0 0.0 1.0
--0.5 4.5 0.800000011920929 0.0 0.0 1.0
-0.5 -4.5 0.800000011920929 0.0 0.0 1.0
-0.5 -3.5 0.800000011920929 0.0 0.0 1.0
-0.5 -2.5 0.800000011920929 0.0 0.0 1.0
-0.5 -1.5000000000000002 0.800000011920929 0.0 0.0 1.0
-0.5000000000000001 -0.49999999999999994 0.800000011920929 0.0 0.0 1.0
-0.5 0.5 0.800000011920929 0.0 0.0 1.0
-0.5000000000000001 1.5 0.800000011920929 0.0 0.0 1.0
-0.5000000000000001 2.5000000000000004 0.800000011920929 0.0 0.0 1.0
-0.5 3.5 0.800000011920929 0.0 0.0 1.0
-0.5 4.5 0.800000011920929 0.0 0.0 1.0
-1.5 -4.5 0.800000011920929 0.0 0.0 1.0
-1.5 -3.5 0.800000011920929 0.0 0.0 1.0
-1.5 -2.5 0.800000011920929 0.0 0.0 1.0
-1.5 -1.5 0.800000011920929 0.0 0.0 1.0
-1.5 -0.5 0.800000011920929 0.0 0.0 1.0
-1.5 0.5 0.800000011920929 0.0 0.0 1.0
-1.5 1.5 0.800000011920929 0.0 0.0 1.0
-1.5 2.5 0.800000011920929 0.0 0.0 1.0
-1.5 3.5 0.800000011920929 0.0 0.0 1.0
-1.5 4.5 0.800000011920929 0.0 0.0 1.0
-2.5 -4.5 0.800000011920929 0.0 0.0 1.0
-2.5 -3.5 0.800000011920929 0.0 0.0 1.0
-2.5 -2.5 0.800000011920929 0.0 0.0 1.0
-2.5 -1.5 0.800000011920929 0.0 0.0 1.0
-2.5 -0.5 0.800000011920929 0.0 0.0 1.0
-2.5 0.5 0.800000011920929 0.0 0.0 1.0
-2.5 1.5 0.800000011920929 0.0 0.0 1.0
-2.5 2.5 0.800000011920929 0.0 0.0 1.0
-2.5 3.5 0.800000011920929 0.0 0.0 1.0
-2.5 4.5 0.800000011920929 0.0 0.0 1.0
-3.5 -4.5 0.800000011920929 0.0 0.0 1.0
-3.5 -3.5 0.800000011920929 0.0 0.0 1.0
-3.5 -2.5 0.800000011920929 0.0 0.0 1.0
-3.5 -1.5 0.800000011920929 0.0 0.0 1.0
-3.5 -0.5 0.800000011920929 0.0 0.0 1.0
-3.5 0.5 0.800000011920929 0.0 0.0 1.0
-3.5 1.5 0.800000011920929 0.0 0.0 1.0
-3.5 2.5 0.800000011920929 0.0 0.0 1.0
-3.5 3.5 0.800000011920929 0.0 0.0 1.0
-3.5 4.5 0.800000011920929 0.0 0.0 1.0
-4.5 -4.5 0.800000011920929 0.0 0.0 1.0
-4.5 -3.5 0.800000011920929 0.0 0.0 1.0
-4.5 -2.5 0.800000011920929 0.0 0.0 1.0
-4.5 -1.5 0.800000011920929 0.0 0.0 1.0
-4.5 -0.5 0.800000011920929 0.0 0.0 1.0
-4.5 0.5 0.800000011920929 0.0 0.0 1.0
-4.5 1.5 0.800000011920929 0.0 0.0 1.0
-4.5 2.5 0.800000011920929 0.0 0.0 1.0
-4.5 3.5 0.800000011920929 0.0 0.0 1.0
-4.5 4.5 0.800000011920929 0.0 0.0 1.0
