--- .ci/unit_tests/python/files/example.res
@@ -1,100 +1,0 @@
-3.07014714
-14.6971494
-6.49928057
-2.56328716
-11.9692467
-12.3453042
-3.66813602
-9.22560809
-20.3026633
-13.6480215
-3.15182842
-5.58012273
-4.77860906
-4.10387572
-5.85612641
-6.40101673
-5.58544798
-8.35932327
-13.6664334
-21.0306921
-2.09158815
-2.69362959
-2.9830171
-3.02886437
-3.60780586
-4.09627538
-4.86106183
-6.98118258
-10.7206036
-17.8391848
-1.30683729
-1.82764012
-2.15424173
-2.31233632
-2.64025716
-3.27558723
-4.18536189
-6.07232798
-9.15007904
-9.44576198
-1.11327099
-1.51548614
-1.79103999
-1.91453567
-2.56990837
-3.01550023
-4.115926
-6.00701625
-9.89075061
-18.6764125
-1.2614028
-1.38964269
-1.5816798
-2.05165146
-2.51459558
-3.09038309
-4.0308652
-5.79225563
-10.1255539
-18.6524802
-1.46894363
-1.83223684
-2.13212628
-2.3274475
-2.66585953
-3.24994369
-4.32989007
-6.0095491
-9.07539487
-9.50361299
-2.06545415
-2.84732973
-3.09566896
-3.0186023
-3.91238868
-4.19165374
-4.84581461
-6.76458363
-10.7765464
-17.7810885
-3.06685354
-5.43427174
-4.70352035
-3.99268629
-5.95401435
-6.48973987
-5.59003575
-8.16623597
-13.7094668
-21.0868981
-2.9564356
-14.8831823
-6.62598214
-2.50453936
-11.8381972
-12.2632417
-3.80144448
-9.46953676
-20.2403713
-13.7475276
