--- .ci/unit_tests/python/honeybee_extension/__init__.py
@@ -1,1 +1,0 @@
-﻿
