--- .ci/unit_tests/python/ladybug_extension/__init__.py
@@ -1,1 +1,0 @@
-﻿
