+++ LadybugTools_Engine/LadybugTools_Engine.csproj
@@ -126,17 +126,17 @@
     <Compile Include="Convert\PascalCaseToSnakeCase.cs" />
     <Compile Include="Convert\SanitiseString.cs" />
     <Compile Include="Convert\SnakeCaseToPascalCase.cs" />
-    <Compile Include="Create\SimulationId.cs" />
+    <Compile Include="Compute\SimulationId.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Query\GetMaterial.cs" />
     <Compile Include="Query\GetTypology.cs" />
-    <Compile Include="Query\LBTLocation.cs" />
     <Compile Include="Query\ToolkitName.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
     <None Include="Python\README.md" />
     <Content Include="Python\requirements.txt" />
+    <None Include="Versioning_60.json" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\LadybugTools_oM\LadybugTools_oM.csproj">
@@ -144,6 +144,9 @@
       <Name>LadybugTools_oM</Name>
       <Private>False</Private>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Create\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\RhinoCommon.5.12.50810.13095\build\net35\RhinoCommon.targets" Condition="Exists('..\packages\RhinoCommon.5.12.50810.13095\build\net35\RhinoCommon.targets')" />
