+++ LadybugTools_oM/Versioning_60.json
@@ -1,0 +1,26 @@
+﻿{
+    "Namespace": {
+        "ToNew": {
+        },
+        "ToOld": {
+        }
+    },
+    "Type": {
+        "ToNew": {
+
+        },
+        "ToOld": {
+        }
+    },
+    "Property": {
+        "ToNew": {
+        },
+        "ToOld": {
+        }
+    },
+    "MessageForDeleted": {
+        "BH.oM.LadybugTools.ExternalComfortMaterial": "Object replaced by OpaqueMaterial and OpaqueVegetationMaterial. If further assistance is needed, please raise an issue on https://github.com/BHoM/LadybugTools_Toolkit/issues."
+    },
+    "MessageForNoUpgrade": {
+    }
+}
